fix dependency path
