+++ x-pack/plugins/siem/public/components/open_timeline/types.ts
@@ -161,7 +161,7 @@ export interface OpenTimelineProps {
   /** the requested field to sort on */
   sortField: string;
   /** timeline / template timeline */
-  tabs: JSX.Element;
+  tabs?: JSX.Element;
   /** The title of the Open Timeline component  */
   title: string;
   /** The total (server-side) count of the search results */
