+++ x-pack/plugins/siem/public/pages/detection_engine/rules/components/step_define_rule/schema.tsx
@@ -168,8 +168,7 @@ export const schema: FormSchema = {
     helpText: i18n.translate(
       'xpack.siem.detectionEngine.createRule.stepAboutRule.fieldTimelineTemplateHelpText',
       {
-        defaultMessage:
-          'Select an existing timeline to use as a template when investigating generated signals.',
+        defaultMessage: 'Select which timeline to use when investigating generated signals.',
       }
     ),
   },
